Initial move of extraction code
